1a
a1

went #2(poop) bathroom finally @ maybe 2:30 til 3 or so 6/17/2020 finally feeling better

air conditioner is on
fan in bedroom on high

A few days ago sodering was done according to dad by people who installed new A/C unit and 

Furnance this set off the smoke alarm in Hallway a few days ago (near big bathroom)

noticed battery missing from smoke alarm (was it missing part of psyop?) [I did take it out to 

stop the beeping but it wasnt there anymore the day or so after]

is now 4:01 AM 6/17/2020

mom earlier around 1:30am maybe 6/17/2020 mentioned fireflies and cold
"cold" reminded me that AC is on and I come in this bedroom and fan is on high
Why is fan on high out of nowhere? WHen AC is on? 

had a thought where house could be rigged with explosives like twin towers
will see if I can check attic area where sodering was done (is sodering required in an attic of 

a house when installing a new A/C Unit and Furnace?). Very suspicio.s